Revert "up #97"
This reverts commit 72117e4ad543d9d117e4eda79fa03d9a33ad27f3.

+++ src/main/java/ps5/takenoko/lanceur/JeuLanceur.java
@@ -113,7 +113,7 @@ public class JeuLanceur {
 
 
         for (int i = 1, j=0; i <= joueurs.size(); i++,j++) {
-            double[] statsRes = stats.getStats(joueurs.get(j), nbparties);
+            Float[] statsRes = stats.getStats(joueurs.get(j), nbparties);
             logs[i]=joueurs.get(j).getClass().getSimpleName()+",";
             for(int k=0;k<statsRes.length;k++){
                 logs[i]+=statsRes[k]+",";
